+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added some processing recipes for TForest in GTC and IC2C machines.
 - Added ability for GT Scanners to be nuclear thermometers as well as EU scanners
 - Changed Beryllium from a fluid to a solid as it should be, recipe to convert from fluid to solid in the Centrifuge.
+- Changed the Dragon Egg Energy Siphon to a Trophy Energy Siphon which can make power from more sources (configurable).
 
 # 1.1.5
 - Added ability to add materials to GTC via Craft Tweaker thanks to Trinsdar.